+++ CHANGELOG.rst
@@ -3,7 +3,14 @@
 Changelog
 =========
 
-Version 0.2.0 - Unreleased
+Version 0.3.0 - Unreleased
+**************************
+
+* Added test for PackageRequirement basic behavior (before the analyzer post process);
+* Implemented Pip inclusion directive (aka ``-r dev.txt``) and added a test for it;
+
+
+Version 0.2.0 - 2024/07/14
 **************************
 
 Added working version of requirement analyzer, parser and package model with